Code/202406/requirements.txt
absl-py==1.4.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
appdirs==1.4.4
cachetools==4.2.4
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
commonmark==0.9.1
crcmod==1.7
cryptography==40.0.2
cycler==0.11.0
Cython==3.0.11
dataclasses==0.8
decorator==5.1.1
distro==1.9.0
fastrlock==0.8.2
google-auth==2.22.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
idna==3.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
jmespath==0.10.0
kiwisolver==1.3.1
Mako==1.1.6
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
model-index==0.1.11
numpy==1.19.5
oauthlib==3.2.2
opendatalab==0.0.2
openmim==0.3.9
ordered-set==4.0.2
oss2==2.18.6
packaging==21.3
pandas==1.1.5
Pillow==6.2.2
platformdirs==2.4.0
protobuf==3.19.6
psutil==6.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pycuda==2020.1
Pygments==2.14.0
PyGObject==3.42.2
pyparsing==3.1.2
PyQt5==5.15.6
PyQt5-sip==12.9.1
python-dateutil==2.9.0.post0
pytools==2022.1.12
pytz==2024.1
PyYAML==6.0.1
requests==2.27.1
requests-oauthlib==2.0.0
rich==12.6.0
rsa==4.9
scikit-build==0.11.1
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch @ file:///home/ubuntu/Downloads/torch-1.10.0-cp36-cp36m-linux_aarch64.whl
torchvision==0.11.0a0+fa347eb
tqdm==4.64.1
typing_extensions==4.1.1
urllib3==1.26.19
watchdog==2.3.1
Werkzeug==2.0.3
zipp==3.6.0
